Move unicode related functions to own module
